+++ test/androidTest/java/org/thoughtcrime/securesms/espresso/HelperSecret.java
@@ -1,0 +1,3 @@
+package org.thoughtcrime.securesms.espresso;
+
+class HelperSecret {};
