+++ test/androidTest/java/org/thoughtcrime/securesms/espresso/PinnedHelper.java
@@ -7,9 +7,8 @@ import static android.support.test.espresso.Espresso.pressBack;
 import static android.support.test.espresso.contrib.RecyclerViewActions.actionOnItemAtPosition;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 
-public class PinnedHelper extends Helper<PinnedHelper> {
-    public PinnedHelper(HelperSecret s) {
-    }
+public class PinnedHelper extends BaseHelper<PinnedHelper> {
+    public PinnedHelper(HelperSecret s) {}
 
     public PinnedHelper unpinMessage(int position) {
         onView(withId(android.R.id.list))
